WARNING! SLIGHT ALTERATION NEEDED!
The game class, which will be the instantiator of other objects, will probably need a small change. I believe that anything that will be passed to the constructor of an object that this class will be making will also need to be passed to that method. So, anything that is needed in the itemClass object constructor needs to be able to be passed to the createItem() method... I think. Since I haven't seen that method yet, I don't know what you put in the constructor and haven't made this change yet.